Added the ability to move through the file system tree via REST requests. Closed all TODOs.

+++ src/main/java/com/vcs/mygit/git/service/impl/FileServiceImpl.java
@@ -1,6 +1,7 @@
 package com.vcs.mygit.git.service.impl;
 
 import com.vcs.mygit.git.dto.RepositoryContext;
+import com.vcs.mygit.git.dto.response.UploadFilesResponse;
 import com.vcs.mygit.git.service.CommandService;
 import com.vcs.mygit.git.service.FileService;
 import lombok.RequiredArgsConstructor;
@@ -15,18 +16,73 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
 @Service
 @RequiredArgsConstructor
-public class FileServiceImpl implements FileService {
+public class FileServiceImpl  implements FileService {
     private final CommandService commandService;
 
-    // TODO: check exception when files have the same name
-    public Map<String, String> uploadFiles(
+    public Object getFileOrDirectoryContents(RepositoryContext repoContext, String path) throws IOException {
+        Path targetPath = repoContext.getRepositoryPath();
+        if (path != null && !path.isBlank()) {
+            targetPath = targetPath.resolve(path);
+        }
+
+        if (Files.isDirectory(targetPath)) {
+            return getDirectoryContents(targetPath);
+        } else if (Files.isRegularFile(targetPath)) {
+            return getFileContent(targetPath);
+        } else {
+            throw new IllegalArgumentException("The path does not exist or is neither a file nor a directory");
+        }
+    }
+
+    public String getFileContent(Path filePath) throws IOException {
+        if (!Files.isRegularFile(filePath)) {
+            throw new IllegalArgumentException("The path does not point to a file");
+        }
+        return Files.readString(filePath);
+    }
+
+    public Map<String, String> getDirectoryContents(Path dirPath) throws IOException {
+        if (!Files.exists(dirPath) || !Files.isDirectory(dirPath)) {
+            throw new IllegalArgumentException("Invalid directory path");
+        }
+
+        Map<String, String> directoryContents = new LinkedHashMap<>();
+        try (Stream<Path> paths = Files.list(dirPath)) {
+            paths.filter(p -> {
+                        try {
+                            return !Files.isHidden(p);
+                        } catch (IOException e) {
+                            throw new RuntimeException(e);
+                        }
+                    }).forEach(p -> {
+                String fileName = p.getFileName().toString();
+                String type = Files.isDirectory(p) ? "directory" : "file";
+                directoryContents.put(fileName, type);
+            });
+        }
+        return directoryContents;
+    }
+
+    private Set<String> getExistingFiles(Path directoryPath) throws IOException {
+        Set<String> existingFiles = new HashSet<>();
+        try (var stream =  Files.walk(directoryPath)) {
+            stream.filter(Files::isRegularFile)
+                    .forEach(file -> {
+                        Path relativePath = directoryPath.relativize(file);
+                        existingFiles.add(relativePath.toString());
+                    });
+        }
+        return existingFiles;
+    }
+
+    public UploadFilesResponse uploadFiles(
             RepositoryContext repoContext,
             MultipartFile[] files
     ) throws IOException, GitAPIException {
@@ -39,6 +95,9 @@ public class FileServiceImpl implements FileService {
             commandService.init(repoContext);
         }
 
+        Set<String> existingFiles = getExistingFiles(repositoryPath);
+
+        List<String> rejectedFiles = new ArrayList<>();
         Map<String, String> addedFiles = new HashMap<>();
         for (MultipartFile file : files) {
             String fileName = file.getOriginalFilename();
@@ -46,14 +105,20 @@ public class FileServiceImpl implements FileService {
             if (fileName == null || fileName.isBlank()) {
                 throw new IllegalArgumentException("The file must have non-blank name");
             }
+
+            if (existingFiles.contains(fileName)) {
+                rejectedFiles.add(fileName);
+                continue;
+            }
+
             Path filePath = repositoryPath.resolve(fileName);
             Files.copy(file.getInputStream(), filePath);
             addedFiles.put(fileName, filePath.toString());
         }
-        return addedFiles;
+        return new UploadFilesResponse(rejectedFiles, addedFiles);
     }
 
-    public void createRepositoryArchive(
+    public void getRepositoryArchive(
             RepositoryContext repoContext,
             HttpServletResponse response
     ) throws IOException {
